Allow to wash_html() with JSON output
output->asset_url() exists only on rcmail_output_html, which made this
method silently depend hard-coded on it.
Now this dependency is moved into the actions that actually work with a
HTML output. This is a tiny bit more verbose, but also much clearer and
more comprehensible.
By default a non-existent URL is used, to prevent washtml keeping remote
image URLs in the HTML (as would happen, without a `blocked_src` key).