Replace inline scripts by JSON-data with instructions
We're now sending JSON-encoded data that instructs the interpreting side
(app.js), which callbacks to call (or events to trigger) with the given
arguments. Basically that means server code now calls $output->command() instead
of add_script().
This way the change is not too radical but we still can get rid of
inline Javascript code.